+++ airframe-http-finagle/src/main/scala/wvlet/airframe/http/finagle/FinagleResponseHandler.scala
@@ -20,7 +20,8 @@ import com.twitter.finagle.http._
 import com.twitter.io.Buf.ByteArray
 import com.twitter.io.{Buf, Reader}
 import wvlet.airframe.codec.{JSONCodec, MessageCodec, MessageCodecFactory}
-import wvlet.airframe.http.{HttpStatus, ResponseHandler, SimpleHttpResponse}
+import wvlet.airframe.http.router.ResponseHandler
+import wvlet.airframe.http.{HttpStatus, SimpleHttpResponse}
 import wvlet.airframe.surface.{Primitive, Surface}
 import wvlet.log.LogSupport
