+++ airframe-http/src/main/scala/wvlet/airframe/http/router/Automaton.scala
@@ -11,8 +11,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.airframe.http
+package wvlet.airframe.http.router
 
+/**
+  *
+  */
 object Automaton {
   def empty[Node, Token]: Automaton[Node, Token] = new Automaton(Set.empty, Set.empty)
