+++ airframe-http/src/main/scala/wvlet/airframe/http/router/ControllerProvider.scala
@@ -11,13 +11,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.airframe.http
+package wvlet.airframe.http.router
 
-import wvlet.log.LogSupport
+import wvlet.airframe._
 import wvlet.airframe.surface.Surface
+import wvlet.log.LogSupport
 
 import scala.util.{Failure, Success, Try}
-import wvlet.airframe._
 
 /**
   *
