+++ airframe-http/src/main/scala/wvlet/airframe/http/router/RouterMacros.scala
@@ -11,9 +11,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.airframe.http
+package wvlet.airframe.http.router
 
-import scala.language.experimental.macros
+import wvlet.airframe.http.HttpFilterType
 import scala.reflect.macros.{blackbox => sm}
 
 /**
