+++ airframe-http/src/test/scala/wvlet/airframe/http/RouterTest.scala
@@ -16,6 +16,7 @@ package wvlet.airframe.http
 import wvlet.airframe.Session
 import wvlet.airframe.http.example.ControllerExample.User
 import wvlet.airframe.http.example._
+import wvlet.airframe.http.router.{ControllerProvider, RouteMatcher}
 import wvlet.airframe.surface.Surface
 import wvlet.airspec.AirSpec
